tools/requirements.txt
requests>=2.19
PyYAML>=3.13
